requirements.txt
pytorch
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
gensim
tqdm
pandas
numpy
scipy
networkx
scikit_learn
torchvision
python-igraph